# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "NyoStore",
    "description" : "This is a NyoStore server. Here you can find \nout all the rest api's.\n",
    "version" : "2.1"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/Developmenttest/NyoStore/2.1",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "http://authlib.nyotek.com:3010",
    "variables" : {
      "protocol" : {
        "default" : "http",
        "enum" : [ "http", "https" ]
      },
      "port" : {
        "default" : "3010",
        "enum" : [ "3010", "3020", "3030", "3040", "3050", "3060", "3070" ]
      }
    }
  } ],
...
  "paths" : {
...
    "/file/{key}" : {
      "get" : {
        "tags" : [ "File" ],
        "summary" : "Get file data",
        "description" : "Returns a single file",
        "operationId" : "getFileByKey",
        "parameters" : [ {
          "name" : "key",
          "in" : "path",
          "description" : "Key of uploaded file to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/File"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX" : {
            "description" : "Unexpected Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "servers" : [ {
        "url" : "http://authlib.nyotek.com:3050",
        "description" : "File service"
      } ]
    },
    "/stream/{key}" : {
      "get" : {
        "tags" : [ "File" ],
        "summary" : "Get stream data",
        "description" : "Returns a single file",
        "operationId" : "getStreamByKey",
        "parameters" : [ {
          "name" : "key",
          "in" : "path",
          "description" : "Key of uploaded file to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "image/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary",
                  "x-content-type" : "image/*"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX" : {
            "description" : "Unexpected Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "servers" : [ {
        "url" : "http://authlib.nyotek.com:3050",
        "description" : "File service"
      } ]
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Error" : {
        "required" : [ "message", "status" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : false
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "Unauthorized" : {
        "required" : [ "message", "status" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : false
          },
          "message" : {
            "type" : "string"
          },
          "info" : {
            "type" : "string"
          }
        }
      },
      "Forbidden" : {
        "required" : [ "message", "status" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : false
          },
          "message" : {
            "type" : "string"
          },
          "info" : {
            "type" : "object"
          }
        }
      },
...
      "File" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "boolean",
            "example" : true
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/File_data"
          }
        },
        "example" : {
          "data" : {
            "owner" : "owner",
            "ext" : "ext",
            "format" : "format",
            "created_at" : "created_at",
            "browser_url" : "browser_url",
            "duration" : "duration",
            "codec" : "codec",
            "filename" : "filename",
            "content_type" : "content_type",
            "size" : 0.8008281904610115,
            "file_id" : "file_id",
            "download_url" : "download_url",
            "checksum" : "checksum",
            "width" : 1.4658129805029452,
            "content_length" : "content_length",
            "height" : 6.027456183070403
          },
          "message" : "message",
          "status" : true
        },
        "xml" : {
          "name" : "File"
        }
      },
...
      "File_data" : {
        "type" : "object",
        "properties" : {
          "file_id" : {
            "type" : "string"
          },
          "owner" : {
            "type" : "string"
          },
          "content_type" : {
            "type" : "string"
          },
          "ext" : {
            "type" : "string"
          },
          "filename" : {
            "type" : "string"
          },
          "browser_url" : {
            "type" : "string"
          },
          "download_url" : {
            "type" : "string"
          },
          "content_length" : {
            "type" : "string"
          },
          "checksum" : {
            "type" : "string"
          },
          "created_at" : {
            "type" : "string"
          },
          "size" : {
            "type" : "number"
          },
          "height" : {
            "type" : "number"
          },
          "width" : {
            "type" : "number"
          },
          "format" : {
            "type" : "string"
          },
          "duration" : {
            "type" : "string"
          },
          "codec" : {
            "type" : "string"
          }
        },
        "example" : {
          "owner" : "owner",
          "ext" : "ext",
          "format" : "format",
          "created_at" : "created_at",
          "browser_url" : "browser_url",
          "duration" : "duration",
          "codec" : "codec",
          "filename" : "filename",
          "content_type" : "content_type",
          "size" : 0.8008281904610115,
          "file_id" : "file_id",
          "download_url" : "download_url",
          "checksum" : "checksum",
          "width" : 1.4658129805029452,
          "content_length" : "content_length",
          "height" : 6.027456183070403
        }
      },
...
    },
...
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}